kiosk-pi/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyserial==3.5
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.19.0
gunicorn==21.2.0
pytest==7.4.3
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0 